Add list handler class
