+++ lib/managers/BookManager.rb
@@ -1,32 +1,26 @@
 require '../handlers/BookListHandler'
+require '../handlers/BookCreatorHandler'
 class BookManager
   def initialize
     @books = []
+    @book = BookCreator.new
+    @list = BookList.new(@books)
   end
 
-  def create_book
+  def add_book
     print 'Title: '
     title = gets.chomp
     print 'Author: '
     author = gets.chomp
-    book = Book.new(title, author)
+    @books << @book.create_book(title, author)
     puts 'The book was added succesfully!'
-    @books << book
   end
 
-  def list_all_books
-    @books.map { |_book| puts book_info }
+  def list_books
+    @list.list_all_books
   end
 
-  def list_book_with_idx
-    @books.map.with_index do |book, i|
-      puts "(#{i + 1}) #{book_info(book)}"
-    end
-  end
-
-  private
-
-  def book_info(book)
-    " - #{book.title} by #{book.author}"
+  def list_books_idx
+    @list.list_book_with_idx
   end
 end
